requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bleach==4.1.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-bootstrap4==21.1
django-clear-cache==0.3
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-summernote==0.8.20.0
django-taggit==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
